requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1 
pika==1.3.1
python-dotenv==1.0.1
mysql==0.0.3
mysql-connector-python
